+++ agrest-engine/src/main/java/io/agrest/runtime/AgCoreModule.java
@@ -2,6 +2,7 @@ package io.agrest.runtime;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.agrest.AgException;
+import io.agrest.access.MaxIncludeDepth;
 import io.agrest.compiler.AgEntityCompiler;
 import io.agrest.compiler.AnnotationsAgEntityCompiler;
 import io.agrest.converter.jsonvalue.Base64Converter;
@@ -127,6 +128,7 @@ import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.util.Date;
 import java.util.Map;
+import java.util.Objects;
 
 /**
  * Configures core Agrest services.
@@ -136,9 +138,11 @@ import java.util.Map;
 public class AgCoreModule implements Module {
 
     private final Map<String, AgEntityOverlay> entityOverlays;
+    private final MaxIncludeDepth maxIncludeDepth;
 
-    protected AgCoreModule(Map<String, AgEntityOverlay> entityOverlays) {
-        this.entityOverlays = entityOverlays;
+    protected AgCoreModule(Map<String, AgEntityOverlay> entityOverlays, MaxIncludeDepth maxIncludeDepth) {
+        this.entityOverlays = Objects.requireNonNull(entityOverlays);
+        this.maxIncludeDepth = Objects.requireNonNull(maxIncludeDepth);
     }
 
     @Override
@@ -147,6 +151,7 @@ public class AgCoreModule implements Module {
         binder.bindList(AgEntityCompiler.class).add(AnnotationsAgEntityCompiler.class);
 
         binder.bindMap(AgEntityOverlay.class).putAll(entityOverlays);
+        binder.bind(MaxIncludeDepth.class).toInstance(maxIncludeDepth);
 
         binder.bindMap(AgExceptionMapper.class)
                 .put(AgException.class.getName(), AgExceptionDefaultMapper.class);